# o_donjon/public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API O'donjon",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://odonjon.fr/",
            "description": "API pour O'donjon"
        }
    ],
    "paths": {
...
        "/login_check": {
            "get": {
                "tags": [
                    "User"
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/connectionUser"
                },
                "responses": {
                    "200": {
                        "description": "data of the connected user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/connectionUser"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Controller\\UserController::browse",
                "responses": {
                    "200": {
                        "description": "List of the users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/userList"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Controller\\UserController::add",
                "requestBody": {
                    "$ref": "#/components/requestBodies/registerUser"
                },
                "responses": {
                    "201": {
                        "description": "Register new user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/user"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Controller\\UserController::read",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User informations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/user"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Controller\\UserController::edit",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/registerUser"
                },
                "responses": {
                    "200": {
                        "description": "User informations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/user"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Controller\\UserController::delete",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/user/campaign/join": {
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Controller\\UserController::joinCampaign",
                "requestBody": {
                    "$ref": "#/components/requestBodies/joinCampaignUser"
                },
                "responses": {
                    "200": {
                        "description": "User informations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Yeah, you have joined the campaign"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Campaign not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This invitation code is not valid !"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "userList": {
                "description": "List of the users",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "7"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "Martin",
                        "nullable": true
                    },
                    "avatarPath": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "avatar_martin.png",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "user": {
                "description": "user informations",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "7"
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 180,
                        "uniqueItems": true,
                        "example": "martin@odonjon.com"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "Martin",
                        "nullable": true
                    },
                    "avatarPath": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "avatar_martin.png",
                        "nullable": true
                    },
                    "campaigns": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/userCampaignList"
                        }
                    },
                    "characters": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/characterList"
                        }
                    }
                },
                "type": "object"
            },
            "userCampaignList": {
                "description": "List of the campaigns",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "7"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "Les poules de ma grand-mère",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "characterList": {
                "description": "List of the characters",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "7"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "Dwalïn",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "connectionUser": {
                "description": "JWT and user data",
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJpYXQiOjE2MjE0OTgwODYsImV4cCI6MTYyMTUyNjg4Niwicm9sZXMiOlsiUk9MRV9VU0VSIl0sInVzZXJuYW1lIjoibmFlbEBvZG9uam9uLmNvbSJ9.tR9dFk5g2tJ1YnV_vGBFmP-37LnaOAqhvgT6VhunNVnmd5xIriR50_hmqa5PMkFFHwSurx2KoxwhpZEOXKabSdRb2yipHszo64TIDNjV35jdli8JAOxX6xSQBqE1lK6-iOKJay7Eeuu1N7TViJdbPfDZodomLDLXOdd3ZHMOb8m3-YerQ0mhHNPf5ze2pYbUgoRG8iOGZJtUnQRYutSqeIEQFqk696gyWJKnG_yd16aPhvzf12gmn4cctOhbpO9-oxt6GDAuMpcEBZVoYhkbNN53QJJ6aCZfp2TpUkbL8io8jPdrUxcil_-yMFVnjBewkzTjFdmypWh2Qo4Dad2K7w"
                    },
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": "7"
                            },
                            "email": {
                                "type": "string",
                                "example": "example@odonjon.com"
                            },
                            "name": {
                                "type": "string",
                                "maxLength": 64,
                                "example": "Martin",
                                "nullable": true
                            },
                            "avatarPath": {
                                "type": "string",
                                "maxLength": 255,
                                "example": "avatar_martin.png",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "notFound": {
                "description": "ressource doesn't existe"
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "ID of the ressource",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "requestBodies": {
            "connectionUser": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "username",
                                "password"
                            ],
                            "properties": {
                                "username": {
                                    "type": "string",
                                    "example": "martin@odonjon.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "password"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "registerUser": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "email",
                                "password",
                                "name",
                                "avatarPath"
                            ],
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "maxLength": 180,
                                    "uniqueItems": true,
                                    "example": "martin@odonjon.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "password"
                                },
                                "name": {
                                    "type": "string",
                                    "maxLength": 64,
                                    "example": "Martin",
                                    "nullable": true
                                },
                                "avatarPath": {
                                    "type": "string",
                                    "maxLength": 255,
                                    "example": "avatar_martin.png",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "joinCampaignUser": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "invitationCode"
                            ],
                            "properties": {
                                "invitationCode": {
                                    "type": "string",
                                    "maxLength": 64,
                                    "example": "60a60a7d4e18a"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            }
        }
    }
}